# vlife-admin/src/main/resources/static/doc/openapi.json
{
  "components": {
    "schemas": {
...
      "Dept": {
        "properties": {
          "id": {
            "format": "string",
            "description": "主键id",
            "type": "string"
          },
          "status": {
            "format": "string",
            "description": "实体状态 【1有效 0逻辑删除】",
            "type": "string"
          },
          "createDate": {
            "format": "string",
            "description": "记录创建日期",
            "type": "string"
          },
          "modifyDate": {
            "format": "string",
            "description": "最近修改日期",
            "type": "string"
          },
          "name": {
            "format": "string",
            "description": "部门名称",
            "type": "string"
          }
        }
      },
      "SysDept": {
        "properties": {
          "id": {
            "format": "string",
            "description": "主键id",
            "type": "string"
          },
          "status": {
            "format": "string",
            "description": "实体状态 【1有效 0逻辑删除】",
            "type": "string"
          },
          "createDate": {
            "format": "string",
            "description": "记录创建日期",
            "type": "string"
          },
          "modifyDate": {
            "format": "string",
            "description": "最近修改日期",
            "type": "string"
          },
          "name": {
            "format": "string",
            "description": "部门名称",
            "type": "string"
          },
          "code": {
            "format": "string",
            "description": "部门编码",
            "type": "string"
          },
          "sysOrgId": {
            "format": "string",
            "description": "机构",
            "type": "string"
          }
        }
      },
...
      "SysArea": {
        "properties": {
          "id": {
            "format": "string",
            "description": "主键id",
            "type": "string"
          },
          "status": {
            "format": "string",
            "description": "实体状态 【1有效 0逻辑删除】",
            "type": "string"
          },
          "createDate": {
            "format": "string",
            "description": "记录创建日期",
            "type": "string"
          },
          "modifyDate": {
            "format": "string",
            "description": "最近修改日期",
            "type": "string"
          },
          "name": {
            "format": "string",
            "description": "区划名称",
            "type": "string"
          },
          "code": {
            "format": "string",
            "description": "区划编码",
            "type": "string"
          },
          "pcode": {
            "format": "string",
            "description": "上级地区编码",
            "type": "string"
          },
          "level": {
            "format": "string",
            "description": "地区类型",
            "type": "string"
          }
        }
      },
...
      "PageVoSysArea": {
        "properties": {
          "result": {
            "description": "No comments found.",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SysArea"
            }
          },
          "size": {
            "format": "int32",
            "description": "No comments found.",
            "type": "integer"
          },
          "page": {
            "format": "int32",
            "description": "No comments found.",
            "type": "integer"
          },
          "total": {
            "format": "int64",
            "description": "No comments found.",
            "type": "number"
          },
          "totalPage": {
            "format": "int64",
            "description": "No comments found.",
            "type": "number"
          },
          "first": {
            "format": "boolean",
            "description": "No comments found.",
            "type": "boolean"
          },
          "last": {
            "format": "boolean",
            "description": "No comments found.",
            "type": "boolean"
          }
        }
      },
...
      "PageVoSysDept": {
        "properties": {
          "result": {
            "description": "No comments found.",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SysDept"
            }
          },
          "size": {
            "format": "int32",
            "description": "No comments found.",
            "type": "integer"
          },
          "page": {
            "format": "int32",
            "description": "No comments found.",
            "type": "integer"
          },
          "total": {
            "format": "int64",
            "description": "No comments found.",
            "type": "number"
          },
          "totalPage": {
            "format": "int64",
            "description": "No comments found.",
            "type": "number"
          },
          "first": {
            "format": "boolean",
            "description": "No comments found.",
            "type": "boolean"
          },
          "last": {
            "format": "boolean",
            "description": "No comments found.",
            "type": "boolean"
          }
        }
      },
...
    }
  },
  "servers": [
    {
      "url": "http://localhost:8088"
    }
  ],
  "openapi": "3.0.3",
  "paths": {
...
    "/sysArea/remove/{id}": {
      "delete": {
        "summary": "逻辑删除;",
        "deprecated": false,
        "description": "逻辑删除;",
        "tags": [
          "行政区划接口;"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "number"
                },
                "examples": {
                  "json": {
                    "summary": "test data",
                    "value": "160"
                  }
                }
              }
            }
          }
        },
        "operationId": "remove",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "description": "主键id;",
            "required": true,
            "example": "144",
            "name": "id"
          }
        ]
      }
    },
...
    "/sysDept/page": {
      "get": {
        "summary": "分页查询null;",
        "deprecated": false,
        "description": "分页查询null;",
        "tags": [
          "null接口;"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PageVoSysDept"
                },
                "examples": {
                  "json": {
                    "summary": "test data",
                    "value": "{\n  \"result\": [\n    {\n      \"id\": \"144\",\n      \"status\": \"tazpkw\",\n      \"createDate\": \"2022-08-13\",\n      \"modifyDate\": \"2022-08-13\",\n      \"name\": \"ellie.kuphal\",\n      \"code\": \"75559\",\n      \"sysOrgId\": \"144\"\n    }\n  ],\n  \"size\": 10,\n  \"page\": 1,\n  \"total\": 361,\n  \"totalPage\": 769,\n  \"first\": true,\n  \"last\": true\n}"
                  }
                }
              }
            }
          }
        },
        "operationId": "page",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "description": "No comments found.",
            "required": false,
            "example": "144",
            "name": "sysOrgId"
          },
          {
            "schema": {
              "type": "array",
              "items": {}
            },
            "in": "query",
            "description": "No comments found.",
            "required": false,
            "example": "9eesbf,54798v",
            "name": "type"
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "description": "No comments found.",
            "required": false,
            "example": "ellie.kuphal",
            "name": "name"
          }
        ]
      }
    },
...
    "/dept/detail/{id}": {
      "get": {
        "summary": "明细查询部门;",
        "deprecated": false,
        "description": "明细查询部门;",
        "tags": [
          "部门接口;"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Dept"
                },
                "examples": {
                  "json": {
                    "summary": "test data",
                    "value": "{\n  \"id\": \"144\",\n  \"status\": \"9obkdq\",\n  \"createDate\": \"2022-08-13\",\n  \"modifyDate\": \"2022-08-13\",\n  \"name\": \"ellie.kuphal\"\n}"
                  }
                }
              }
            }
          }
        },
        "operationId": "detail",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "description": "主键id;",
            "required": true,
            "example": "144",
            "name": "id"
          }
        ]
      }
    },
    "/dept/save": {
      "post": {
        "summary": "保存部门;",
        "deprecated": false,
        "description": "保存部门;",
        "tags": [
          "部门接口;"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Dept"
              },
              "examples": {
                "json": {
                  "summary": "test data",
                  "value": "{\n  \"id\": \"144\",\n  \"status\": \"xzgwo2\",\n  \"createDate\": \"2022-08-13\",\n  \"modifyDate\": \"2022-08-13\",\n  \"name\": \"ellie.kuphal\"\n}"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Dept"
                },
                "examples": {
                  "json": {
                    "summary": "test data",
                    "value": "{\n  \"id\": \"144\",\n  \"status\": \"clpsie\",\n  \"createDate\": \"2022-08-13\",\n  \"modifyDate\": \"2022-08-13\",\n  \"name\": \"ellie.kuphal\"\n}"
                  }
                }
              }
            }
          }
        },
        "operationId": "save",
        "parameters": []
      }
    },
...
    "/sysArea/page": {
      "get": {
        "summary": "分页查询行政区划;",
        "deprecated": false,
        "description": "分页查询行政区划;",
        "tags": [
          "行政区划接口;"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PageVoSysArea"
                },
                "examples": {
                  "json": {
                    "summary": "test data",
                    "value": "{\n  \"result\": [\n    {\n      \"id\": \"144\",\n      \"status\": \"v0gchx\",\n      \"createDate\": \"2022-08-13\",\n      \"modifyDate\": \"2022-08-13\",\n      \"name\": \"ellie.kuphal\",\n      \"code\": \"75559\",\n      \"pcode\": \"75559\",\n      \"level\": \"0mgbei\"\n    }\n  ],\n  \"size\": 10,\n  \"page\": 1,\n  \"total\": 20,\n  \"totalPage\": 797,\n  \"first\": true,\n  \"last\": true\n}"
                  }
                }
              }
            }
          }
        },
        "operationId": "page",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "description": "No comments found.",
            "required": false,
            "example": "ellie.kuphal",
            "name": "name"
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "description": "No comments found.",
            "required": false,
            "example": "o441m6",
            "name": "level"
          }
        ]
      }
    },
...
    "/sysDept/save": {
      "post": {
        "summary": "保存null;",
        "deprecated": false,
        "description": "保存null;",
        "tags": [
          "null接口;"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SysDept"
              },
              "examples": {
                "json": {
                  "summary": "test data",
                  "value": "{\n  \"id\": \"144\",\n  \"status\": \"n9w5hz\",\n  \"createDate\": \"2022-08-13\",\n  \"modifyDate\": \"2022-08-13\",\n  \"name\": \"ellie.kuphal\",\n  \"code\": \"75559\",\n  \"sysOrgId\": \"144\"\n}"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/SysDept"
                },
                "examples": {
                  "json": {
                    "summary": "test data",
                    "value": "{\n  \"id\": \"144\",\n  \"status\": \"yqekpi\",\n  \"createDate\": \"2022-08-13\",\n  \"modifyDate\": \"2022-08-13\",\n  \"name\": \"ellie.kuphal\",\n  \"code\": \"75559\",\n  \"sysOrgId\": \"144\"\n}"
                  }
                }
              }
            }
          }
        },
        "operationId": "save",
        "parameters": []
      }
    },
...
    "/sysDept/remove/{id}": {
      "delete": {
        "summary": "逻辑删除;",
        "deprecated": false,
        "description": "逻辑删除;",
        "tags": [
          "null接口;"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "number"
                },
                "examples": {
                  "json": {
                    "summary": "test data",
                    "value": "365"
                  }
                }
              }
            }
          }
        },
        "operationId": "remove",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "description": "主键id;",
            "required": true,
            "example": "144",
            "name": "id"
          }
        ]
      }
    },
...
    "/sysArea/detail/{id}": {
      "get": {
        "summary": "明细查询行政区划;",
        "deprecated": false,
        "description": "明细查询行政区划;",
        "tags": [
          "行政区划接口;"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/SysArea"
                },
                "examples": {
                  "json": {
                    "summary": "test data",
                    "value": "{\n  \"id\": \"144\",\n  \"status\": \"2bs5dh\",\n  \"createDate\": \"2022-08-13\",\n  \"modifyDate\": \"2022-08-13\",\n  \"name\": \"ellie.kuphal\",\n  \"code\": \"75559\",\n  \"pcode\": \"75559\",\n  \"level\": \"43p9kt\"\n}"
                  }
                }
              }
            }
          }
        },
        "operationId": "detail",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "description": "主键id;",
            "required": true,
            "example": "144",
            "name": "id"
          }
        ]
      }
    },
...
    "/sysArea/save": {
      "post": {
        "summary": "保存行政区划;",
        "deprecated": false,
        "description": "保存行政区划;",
        "tags": [
          "行政区划接口;"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SysArea"
              },
              "examples": {
                "json": {
                  "summary": "test data",
                  "value": "{\n  \"id\": \"144\",\n  \"status\": \"iulg8m\",\n  \"createDate\": \"2022-08-13\",\n  \"modifyDate\": \"2022-08-13\",\n  \"name\": \"ellie.kuphal\",\n  \"code\": \"75559\",\n  \"pcode\": \"75559\",\n  \"level\": \"3xd5jv\"\n}"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/SysArea"
                },
                "examples": {
                  "json": {
                    "summary": "test data",
                    "value": "{\n  \"id\": \"144\",\n  \"status\": \"9r9shf\",\n  \"createDate\": \"2022-08-13\",\n  \"modifyDate\": \"2022-08-13\",\n  \"name\": \"ellie.kuphal\",\n  \"code\": \"75559\",\n  \"pcode\": \"75559\",\n  \"level\": \"woluny\"\n}"
                  }
                }
              }
            }
          }
        },
        "operationId": "save",
        "parameters": []
      }
    },
...
    "/dept/remove/{id}": {
      "delete": {
        "summary": "逻辑删除;",
        "deprecated": false,
        "description": "逻辑删除;",
        "tags": [
          "部门接口;"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "number"
                },
                "examples": {
                  "json": {
                    "summary": "test data",
                    "value": "560"
                  }
                }
              }
            }
          }
        },
        "operationId": "remove",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "description": "主键id;",
            "required": true,
            "example": "144",
            "name": "id"
          }
        ]
      }
    },
...
    "/sysDept/detail/{id}": {
      "get": {
        "summary": "明细查询null;",
        "deprecated": false,
        "description": "明细查询null;",
        "tags": [
          "null接口;"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/SysDept"
                },
                "examples": {
                  "json": {
                    "summary": "test data",
                    "value": "{\n  \"id\": \"144\",\n  \"status\": \"ipd9s1\",\n  \"createDate\": \"2022-08-13\",\n  \"modifyDate\": \"2022-08-13\",\n  \"name\": \"ellie.kuphal\",\n  \"code\": \"75559\",\n  \"sysOrgId\": \"144\"\n}"
                  }
                }
              }
            }
          }
        },
        "operationId": "detail",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "description": "主键id;",
            "required": true,
            "example": "144",
            "name": "id"
          }
        ]
      }
    }
  },
  "info": {
    "title": "test",
    "version": "1.0.0"
  },
...
}